refactor: add policy folder
