+++ scripts/s2n_safety_explicit_context/codemod.sh
@@ -27,6 +27,7 @@ IGNORES=$(
         docker-images \
         docs \
         error \
+        policy \
         lib \
         libcrypto-build \
         libcrypto-root \
